+++ fem/integ/lininteg_domain_grad.cpp
@@ -9,9 +9,9 @@
 // terms of the BSD-3 license. We welcome feedback and contributions, see file
 // CONTRIBUTING.md for details.
 
-#include "fem.hpp"
-#include "../fem/kernels.hpp"
-#include "../general/forall.hpp"
+#include "../../general/forall.hpp"
+#include "../../fem/kernels.hpp"
+#include "../fem.hpp"
 
 namespace mfem
 {
